requirements.txt
pandas
requests
beautifulsoup4
numpy
tqdm
playwright
python-dotenv
urllib3
